changed Landing to Home and Home to Tasks; changed routes and imports to align w/ changes
