+++ public/css/home_auth.css
@@ -314,7 +314,7 @@ canvas {
 }
 
 .container.signup-signin {
-    padding-top: 10%;
+    padding-top: 7%;
     width: 60%;
 }
