+++ resources/js/components/layouts/HomeNavbar.js
@@ -1,14 +1,22 @@
 import React from "react";
 import {Link} from 'react-router-dom';
 
-const HomeNavbar = () => {
+const HomeNavbar = ({isLoggedIn}) => {
+
     return (
         <nav id="navbar" className="navbar order-last order-lg-0">
             <ul>
-                <li><Link className="active" to="/home">Tasks</Link></li>
-                <li><Link to="">Inventory</Link></li>
-                <li><Link to="">Shop</Link></li>
+                <li><Link className="active" to="/">Home</Link></li>
+                <li><Link to="">About</Link></li>
                 <li><Link to="">Features</Link></li>
+                <li><Link to="">FAQs</Link></li>
+                <li><Link to="">Pricing</Link></li>
+                <li><Link to="">Contact</Link></li>
+                {
+                    isLoggedIn === "true"
+                    ? <li><Link to="/tasks">My Tasks</Link></li>
+                    : null
+                }
             </ul>
             <i className="bi bi-list mobile-nav-toggle"></i>
         </nav>
@@ -16,3 +24,4 @@ const HomeNavbar = () => {
 }
 
 export default HomeNavbar;
+
