+++ resources/js/components/layouts/TasksNavbar.js
@@ -1,0 +1,18 @@
+import React from "react";
+import {Link} from 'react-router-dom';
+
+const TasksNavbar = () => {
+    return (
+        <nav id="navbar" className="navbar order-last order-lg-0">
+            <ul>
+                <li><Link className="active" to="/tasks">Tasks</Link></li>
+                <li><Link to="">Inventory</Link></li>
+                <li><Link to="">Shop</Link></li>
+                <li><Link to="">Party</Link></li>
+            </ul>
+            <i className="bi bi-list mobile-nav-toggle"></i>
+        </nav>
+    );
+}
+
+export default TasksNavbar;
