+++ resources/js/components/routes/AppRouter.js
@@ -3,10 +3,10 @@ import ReactDOM from 'react-dom';
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
 import axios from 'axios';
 
-import Landing from "../pages/Landing";
 import Auth from "../pages/Auth";
 import Shop from "../pages/shop/shop";
 import Home from "../pages/Home";
+import Tasks from "../pages/Tasks";
 
 axios.defaults.baseURL = "http://127.0.0.1:8000/";
 axios.defaults.headers.post['Content-Type'] = 'application/json';
@@ -24,10 +24,10 @@ const AppRouter = () => {
         <div>
             <Router>
                 <Routes>
-                    <Route path='/' exact element={<Landing/>}></Route>
+                    <Route path='/' exact element={<Home/>}></Route>
                     <Route path='/register' element={<Auth page="register"/>}></Route>
                     <Route path='/login' element={<Auth page="login"/>}></Route>
-                    <Route path='/home' element={<Home/>}></Route>
+                    <Route path='/tasks' element={<Tasks/>}></Route>
                     <Route path='/shop' element={<Shop/>}></Route>
                 </Routes>
             </Router>
